refactor(vjsc): delegate output formatting
