+++ packages/skins/package.json
@@ -38,6 +38,7 @@
     "dashjs": "^5.2.0",
     "hls.js": "^1.6.7",
     "mux-embed": "5.17.10",
+    "oxfmt": "0.61.0",
     "react": "^19.2.8",
     "react-dom": "^19.2.8",
     "shadcn": "^4.19.0",
