--- packages/vjsc/src/output/index.ts
@@ -1,6 +1,0 @@
-export {
-  createOxfmtSourceFormatter,
-  type EditableSource,
-  type OxfmtSourceFormatterOptions,
-  type SourceFormatter,
-} from './format';
